# api/openapi.yaml
openapi: 3.0.3
info:
  title: Matrix Runtime API
  description: |
    Execution plane for Matrix Cloud. Runs MCP sandboxes, inspects Hugging Face
    models and (in future) agents/tools. Sandboxes are a thin alias over jobs of
    type mcp.test.
  version: 0.2.0
servers:
  - url: http://localhost:8080
    description: Local
security:
  - bearerAuth: []
  - {}
paths:
...
  /v1/jobs:
    get:
      summary: List jobs (newest first)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobs:
                    type: array
                    items: { $ref: "#/components/schemas/Job" }
    post:
      summary: Create a job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobRequest"
      responses:
        "202":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateJobResponse"
        "400": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
...
  /v1/jobs/{job_id}/events:
    parameters:
      - $ref: "#/components/parameters/JobID"
    get:
      summary: Stream job events (SSE)
      responses:
        "200":
          description: text/event-stream of Event objects
          content:
            text/event-stream:
              schema:
                $ref: "#/components/schemas/Event"
  /v1/sandbox/sessions:
    post:
      summary: Create a sandbox session (alias over mcp.test)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SandboxRequest"
      responses:
        "202":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SandboxResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    JobID:
      name: job_id
      in: path
      required: true
      schema: { type: string }
...
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
              status: { type: integer }
  schemas:
...
    CreateJobRequest:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [mcp.test, mcp.run, model.inspect, model.pull, model.preload, agent.run, tool.run]
        ttl_seconds: { type: integer, example: 600 }
        payload:
          type: object
          additionalProperties: true
    CreateJobResponse:
      type: object
      properties:
        job_id: { type: string }
        status: { type: string }
        events_url: { type: string }
    Job:
      type: object
      properties:
        job_id: { type: string }
        type: { type: string }
        status:
          type: string
          enum: [queued, running, complete, error, expired, cancelled]
        created_at: { type: string, format: date-time }
        expires_at: { type: string, format: date-time }
        result: { nullable: true }
        error: { type: string }
    Event:
      type: object
      properties:
        step: { type: string }
        status:
          type: string
          enum: [queued, start, ok, error, running, expired, cancelled, complete]
        message: { type: string }
        data: { type: object, additionalProperties: true }
    SandboxRequest:
      type: object
      required: [start_command]
      properties:
        entity_id: { type: string }
        ttl_seconds: { type: integer }
        runtime: { type: string, example: node }
        transport: { type: string, example: stdio }
        start_command: { type: string }
        env:
          type: object
          additionalProperties: { type: string }
    SandboxResponse:
      type: object
      properties:
        session_id: { type: string }
        job_id: { type: string }
        status: { type: string }
        expires_at: { type: string, format: date-time }
        events_url: { type: string }
...